backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
minio==7.1.16
pymongo==4.5.0
boto3==1.28.62
Pillow==10.0.1
pytesseract==0.3.10
PyPDF2==3.0.1
groq==0.3.0
openai==0.28.1
python-dotenv==1.0.0